chore: move constants into new file
