+++ packages/ui/src/1_atoms/Link/Link.tsx
@@ -2,7 +2,7 @@ import React, { FC, PropsWithChildren, ReactNode } from 'react';
 
 import classNames from 'classnames';
 
-import { DATA_ATTRIBUTE } from '../../types';
+import { DATA_ATTRIBUTE } from '../../utils/constants';
 import styles from './Link.module.css';
 import { LinkStyle } from './Link.types';
