+++ packages/ui/src/1_atoms/Paragraph/Paragraph.tsx
@@ -2,7 +2,7 @@ import React, { ReactNode } from 'react';
 
 import classNames from 'classnames';
 
-import { DATA_ATTRIBUTE } from '../../types';
+import { DATA_ATTRIBUTE } from '../../utils/constants';
 import styles from './Paragraph.module.css';
 import { ParagraphSize, ParagraphStyle } from './Paragraph.types';
