+++ packages/ui/src/2_molecules/Dialog/Dialog.tsx
@@ -3,7 +3,7 @@ import React, { useCallback, useEffect, MouseEvent, useRef } from 'react';
 import classNames from 'classnames';
 import FocusTrap from 'focus-trap-react';
 
-import { DATA_ATTRIBUTE } from '../../types';
+import { DATA_ATTRIBUTE } from '../../utils/constants';
 import { Overlay, OverlayProps } from '../Overlay/Overlay';
 import styles from './Dialog.module.css';
 import {
