+++ packages/ui/src/2_molecules/Dropdown/Dropdown.tsx
@@ -16,7 +16,7 @@ import { IconNames } from '../../1_atoms/Icon/Icon.types';
 import { Portal } from '../../1_atoms/Portal/Portal';
 import { useOnClickOutside } from '../../hooks/useOnClickOutside';
 import { Nullable } from '../../types';
-import { DATA_ATTRIBUTE } from '../../types';
+import { DATA_ATTRIBUTE } from '../../utils/constants';
 import styles from './Dropdown.module.css';
 import { DropdownCoords, DropdownMode, DropdownSize } from './Dropdown.types';
 import { getDropdownPositionStyles } from './Dropdown.utils';
