+++ packages/ui/src/2_molecules/SimpleTable/SimpleTable.tsx
@@ -2,7 +2,7 @@ import React, { FC, PropsWithChildren } from 'react';
 
 import classNames from 'classnames';
 
-import { DATA_ATTRIBUTE } from '../../types';
+import { DATA_ATTRIBUTE } from '../../utils/constants';
 import styles from './SimpleTable.module.css';
 
 type SimpleTableProps = {
