+++ packages/ui/src/2_molecules/SimpleTable/components/SimpleTableRow/SimpleTableRow.tsx
@@ -2,7 +2,7 @@ import React, { FC, ReactNode } from 'react';
 
 import classNames from 'classnames';
 
-import { DATA_ATTRIBUTE } from '../../../../types';
+import { DATA_ATTRIBUTE } from '../../../../utils/constants';
 import styles from './SimpleTableRow.module.css';
 
 type SimpleTableRowProps = {
