+++ packages/ui/src/2_molecules/Tooltip/Tooltip.tsx
@@ -15,8 +15,8 @@ import classNames from 'classnames';
 
 import { Portal } from '../../1_atoms/Portal/Portal';
 import { useOnClickOutside } from '../../hooks/useOnClickOutside';
-import { DATA_ATTRIBUTE, Nullable } from '../../types';
 import { noop } from '../../utils';
+import { DATA_ATTRIBUTE, Nullable } from '../../utils/constants';
 import styles from './Tooltip.module.css';
 import {
   TooltipElements,
