+++ packages/ui/src/2_molecules/TransactionId/TransactionId.tsx
@@ -5,8 +5,8 @@ import classNames from 'classnames';
 import { Icon, Link } from '../../1_atoms';
 import { IconNames } from '../../1_atoms/Icon/Icon.types';
 import { LinkStyle } from '../../1_atoms/Link/Link.types';
-import { DATA_ATTRIBUTE } from '../../types';
 import { prettyTx } from '../../utils';
+import { DATA_ATTRIBUTE } from '../../utils/constants';
 import { Tooltip } from '../Tooltip';
 import { TooltipPlacement, TooltipTrigger } from '../Tooltip/Tooltip.types';
 import styles from './TransactionId.module.css';
