+++ packages/ui/src/types/index.ts
@@ -1,5 +1,3 @@
 export * from './tailwind';
 
 export type Nullable<T = any> = T | null;
-
-export const DATA_ATTRIBUTE = 'data-layout-id';
